Move sources list to the bottom and fix its styling
